Move iterators to their own file
